Week3/may 22/requirements.txt
google-generativeai>=0.3.1
langchain>=0.1.0
langchain-google-genai>=0.0.5
chromadb>=0.4.18
pydantic>=2.5.2
asyncio>=3.4.3
python-dotenv>=1.0.0
langchain-community>=0.0.13
langchain-core>=0.1.8
